requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.26.0
ta>=0.10.2
python-dotenv>=0.19.0
langchain>=0.1.0
openai>=1.0.0
